requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4~=4.10.0
billiard==3.6.4.0
bs4~=0.0.1
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-filter==2.4.0
django-ratelimit==3.0.1
djangorestframework==3.12.4
ffmpeg-python==0.2.0
future==0.18.2
idna==3.2
instalooter~=2.4.4
jdatetime==3.6.2
kavenegar==1.1.2
kombu==5.1.0
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-decouple==3.4
pytube~=11.0.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
